CI 4 Changes - Corrected improper formatting - Upgraded form validation to CI4 style - Corrected PHPdoc - Removed left/right hand statements that were equal to each other - Corrected function signatures to indicate possible return types
