+++ app/Views/errors/error.php
@@ -1,0 +1,1 @@
+<div class="error"><?= esc($error) ?></div>
